refactor(execution): remove unused container engine seam
